Move userRoleTable into rbac.ts
