+++ src/data/schema/index.ts
@@ -3,5 +3,4 @@ export * from './rbac'
 export * from './refresh-tokens'
 export * from './teams'
 export * from './tokens'
-export * from './user-role'
 export * from './users'
