+++ src/data/schema/rbac.ts
@@ -3,6 +3,7 @@ import {
   pgTable,
   primaryKey,
   serial,
+  timestamp,
   uniqueIndex,
   uuid,
 } from 'drizzle-orm/pg-core'
@@ -15,6 +16,13 @@ import { usersTable } from './users'
 export const roleEnum = createPgEnum('role', Role)
 export const actionEnum = createPgEnum('action', Action)
 export const resourceEnum = createPgEnum('resource', Resource)
+
+export const userRoleTable = pgTable('user_role', {
+  userId: uuid('user_id').primaryKey().references(() => usersTable.id, { onDelete: 'cascade' }),
+  role: roleEnum('role').notNull().$type<Role>(),
+  invitedBy: uuid('invited_by').references(() => usersTable.id, { onDelete: 'set null' }),
+  assignedAt: timestamp('assigned_at', { withTimezone: true }).notNull().defaultNow(),
+})
 
 export const permissionsTable = pgTable('permissions', {
   id: serial('id').primaryKey(),
